Add log funcs.
